# Optimization for globpath (root elements for recursion)