--- a/internal/globpath/globpath.go
+++ b/internal/globpath/globpath.go
@@ -16,7 +16,6 @@ type GlobPath struct {
 	hasMeta      bool
 	hasSuperMeta bool
 	g            glob.Glob
-	root         string
 }
 
 func Compile(path string) (*GlobPath, error) {
@@ -27,7 +26,7 @@ func Compile(path string) (*GlobPath, error) {
 	}
 
 	// if there are no glob meta characters in the path, don't bother compiling
-	// a glob object or finding the root directory. (see short-circuit in Match)
+	// a glob object
 	if !out.hasMeta || !out.hasSuperMeta {
 		return &out, nil
 	}
@@ -36,14 +35,12 @@ func Compile(path string) (*GlobPath, error) {
 	if out.g, err = glob.Compile(path, os.PathSeparator); err != nil {
 		return nil, err
 	}
-	// Get the root directory for this filepath
-	out.root = findRootDir(path)
 	return &out, nil
 }
 
 func (g *GlobPath) Match() map[string]os.FileInfo {
+	out := make(map[string]os.FileInfo)
 	if !g.hasMeta {
-		out := make(map[string]os.FileInfo)
 		info, err := os.Stat(g.path)
 		if err == nil {
 			out[g.path] = info
@@ -51,7 +48,6 @@ func (g *GlobPath) Match() map[string]os.FileInfo {
 		return out
 	}
 	if !g.hasSuperMeta {
-		out := make(map[string]os.FileInfo)
 		files, _ := filepath.Glob(g.path)
 		for _, file := range files {
 			info, err := os.Stat(file)
@@ -61,48 +57,37 @@ func (g *GlobPath) Match() map[string]os.FileInfo {
 		}
 		return out
 	}
-	return walkFilePath(g.root, g.g)
-}
-
-// walk the filepath from the given root and return a list of files that match
-// the given glob.
-func walkFilePath(root string, g glob.Glob) map[string]os.FileInfo {
-	matchedFiles := make(map[string]os.FileInfo)
+	roots, err := g.findRoots()
+	if err != nil {
+		return out
+	}
 	walkfn := func(path string, info os.FileInfo, _ error) error {
-		if g.Match(path) {
-			matchedFiles[path] = info
+		if g.g.Match(path) {
+			out[path] = info
 		}
 		return nil
-	}
-	filepath.Walk(root, walkfn)
-	return matchedFiles
-}
 
-// find the root dir of the given path (could include globs).
-// ie:
-//   /var/log/telegraf.conf -> /var/log
-//   /home/** ->               /home
-//   /home/*/** ->             /home
-//   /lib/share/*/*/**.txt ->  /lib/share
-func findRootDir(path string) string {
-	pathItems := strings.Split(path, sepStr)
-	out := sepStr
-	for i, item := range pathItems {
-		if i == len(pathItems)-1 {
-			break
-		}
-		if item == "" {
-			continue
-		}
-		if hasMeta(item) {
-			break
-		}
-		out += item + sepStr
 	}
-	if out != "/" {
-		out = strings.TrimSuffix(out, "/")
+	for _, root := range roots {
+		filepath.Walk(root, walkfn)
 	}
 	return out
+}
+
+// find the root elements of the object path, the entry point for recursion
+// when you have a super-meta in your path (which are : 
+// glob(/your/expression/until/first/star/of/super-meta))
+// ie:
+//   /var/log/telegraf.conf -> /var/log/telegraf.conf
+//   /home/** ->               filepath.Glob(/home/*)
+//   /home/*/** ->             filepath.Glob(/home/*/*)
+//   /lib/share/*/*/**.txt ->  filepath.Glob(/lib/share/*/*/*)
+func (g *GlobPath) findRoots() ([]string, error) {
+	if !g.hasSuperMeta {
+		return filepath.Glob(g.path)
+	}
+	rootGlob := g.path[:strings.Index(g.path, "**")+1]
+	return filepath.Glob(rootGlob)
 }
 
 // hasMeta reports whether path contains any magic glob characters.