When you have a super-meta in your expression (`**`), globpath was trying every files descendant of a root directory which was defined as « truncate path at the last / before the first star ». In cases such as `/path/**` or `/some/where/*/**`, it was OK.

But for path like `/some/*/path/*with/stars*/and/elements**` it was not optimal : it recursed through all directories under '/some', while it's only necessary to recurse through directories under `Glob("/some/*/path/*with/stars*/and/elements*")`.

This new version of globpath defines 'roots' (files and directories) as the filepath.Glob (« truncate between the 2 stars of the first `**` occurence »), and then calls filepath.Walk from these roots.

Simple tests (using 'time' on unix), where telegraf contains telegraf source code, and calling filestat agains these paths (md5=false) :

| | `telegraf/plugins/aggregators/**` | `telegraf/*lugins/aggregators/**`|
| -|-|-|
| master|real    0m0,093s|real    0m0,510s|
| optimized version|real    0m0,091s|real    0m0,088s|

(the differences between 0,088 and 0,093s are non significant I think).

I also tried using [godirwalk](https://github.com/karrick/godirwalk) as suggested by @danielnelson in #4778 but the improvement in performance was about 6% for recursing through telegraf source tree (with unsorted=true). Is it worth adding a dependency ?

And finally i moved from 'assert' to 'require' for some tests in globpath.go, as asked in [wiki/Review](https://github.com/influxdata/telegraf/wiki/Review), but i'm not sure if it's a good idea (I leave to you, maintainers, to decide, that's just 2219c2a commit to include or not).

### Required for all PRs:

- [x] Signed [CLA](https://influxdata.com/community/cla/).
- [x] (no need) Associated README.md updated.
- [x] Has appropriate unit tests.